+++ src/emu/CMakeLists.txt
@@ -83,12 +83,30 @@ elseif(EMSCRIPTEN)
     target_sources(emu_core PRIVATE
         ${RP6502_SRC}/emu/host/posix/dir.c
         ${RP6502_SRC}/emu/host/web/fs.c
-        ${RP6502_SRC}/emu/host/posix/os.c)
-else()
+        ${RP6502_SRC}/emu/host/posix/os.c
+        ${RP6502_SRC}/emu/host/web/os.c)
+elseif(ANDROID)
+    # Bionic has no <aio.h> and we don't thread host I/O, so Android gets its own
+    # synchronous fs seam (like web). dir stays on posix; os = posix common + android.
+    target_sources(emu_core PRIVATE
+        ${RP6502_SRC}/emu/host/posix/dir.c
+        ${RP6502_SRC}/emu/host/android/fs.c
+        ${RP6502_SRC}/emu/host/posix/os.c
+        ${RP6502_SRC}/emu/host/android/os.c)
+elseif(APPLE)
+    # macOS: POSIX aio_read is in libc (no librt needed); os = posix common + macos.
     target_sources(emu_core PRIVATE
         ${RP6502_SRC}/emu/host/posix/dir.c
         ${RP6502_SRC}/emu/host/posix/fs.c
-        ${RP6502_SRC}/emu/host/posix/os.c)
+        ${RP6502_SRC}/emu/host/posix/os.c
+        ${RP6502_SRC}/emu/host/macos/os.c)
+else()
+    # Linux / BSD: os = posix common + linux; POSIX AIO required.
+    target_sources(emu_core PRIVATE
+        ${RP6502_SRC}/emu/host/posix/dir.c
+        ${RP6502_SRC}/emu/host/posix/fs.c
+        ${RP6502_SRC}/emu/host/posix/os.c
+        ${RP6502_SRC}/emu/host/linux/os.c)
     include(CheckSymbolExists)
     find_library(RT_LIBRARY rt)
     if(RT_LIBRARY)
@@ -96,11 +114,7 @@ else()
         set(CMAKE_REQUIRED_LIBRARIES ${RT_LIBRARY})
     endif()
     set(CMAKE_REQUIRED_DEFINITIONS -D_GNU_SOURCE)
-    if(ANDROID)
-        set(EMU_POSIX_AIO ON)
-    else()
-        check_symbol_exists(aio_read "aio.h" EMU_POSIX_AIO)
-    endif()
+    check_symbol_exists(aio_read "aio.h" EMU_POSIX_AIO)
     unset(CMAKE_REQUIRED_DEFINITIONS)
     unset(CMAKE_REQUIRED_LIBRARIES)
     if(NOT EMU_POSIX_AIO)
@@ -202,9 +216,31 @@ if(EMU_HAS_WINDOW)
 endif()
 
 # Build source list with window and audio options
-set(APP_SOURCES ${RP6502_SRC}/emu/app/main.c ${RP6502_SRC}/emu/app/cli.c ${RP6502_SRC}/emu/app/window.c ${RP6502_SRC}/emu/app/input.c ${RP6502_SRC}/emu/app/png.c ${RP6502_SRC}/emu/app/web_exports.c)
+set(APP_SOURCES ${RP6502_SRC}/emu/app/cli.c ${RP6502_SRC}/emu/app/input.c ${RP6502_SRC}/emu/app/png.c)
+if(NOT ANDROID)
+    # Desktop/web enter through main(); Android enters through host/android sokol_main.
+    list(APPEND APP_SOURCES ${RP6502_SRC}/emu/app/main.c)
+endif()
+if(EMSCRIPTEN)
+    list(APPEND APP_SOURCES ${RP6502_SRC}/emu/host/web/exports.c)
+endif()
+# Per-host window: the shared render core + exactly one host's window, or the
+# headless stub when there is no window backend.
 if(EMU_HAS_WINDOW)
-    list(APPEND APP_SOURCES ${RP6502_SRC}/emu/app/sokol_impl.c)
+    list(APPEND APP_SOURCES ${RP6502_SRC}/emu/app/sokol_impl.c ${RP6502_SRC}/emu/host/window_core.c)
+    if(WIN32)
+        list(APPEND APP_SOURCES ${RP6502_SRC}/emu/host/win/window.c)
+    elseif(APPLE)
+        list(APPEND APP_SOURCES ${RP6502_SRC}/emu/host/macos/window.c)
+    elseif(EMSCRIPTEN)
+        list(APPEND APP_SOURCES ${RP6502_SRC}/emu/host/web/window.c)
+    elseif(ANDROID)
+        list(APPEND APP_SOURCES ${RP6502_SRC}/emu/host/android/window.c)
+    else()
+        list(APPEND APP_SOURCES ${RP6502_SRC}/emu/host/linux/window.c)
+    endif()
+else()
+    list(APPEND APP_SOURCES ${RP6502_SRC}/emu/host/window_headless.c)
 endif()
 if(EMU_HAS_AUDIO)
     list(APPEND APP_SOURCES ${RP6502_SRC}/emu/app/audio_out.c)
@@ -248,7 +284,7 @@ else()
         # sokol on macOS needs Objective-C; compile the impl units as .m and
         # link the Cocoa/Metal frameworks. (macOS bring-up pass.)
         target_compile_definitions(rp6502-emu PRIVATE SOKOL_METAL)
-        set_source_files_properties(${RP6502_SRC}/emu/app/sokol_impl.c ${RP6502_SRC}/emu/app/window.c ${RP6502_SRC}/emu/app/input.c
+        set_source_files_properties(${RP6502_SRC}/emu/app/sokol_impl.c ${RP6502_SRC}/emu/host/window_core.c ${RP6502_SRC}/emu/host/macos/window.c ${RP6502_SRC}/emu/app/input.c
             PROPERTIES COMPILE_FLAGS "-x objective-c")
         target_link_libraries(rp6502-emu PRIVATE
             "-framework Cocoa" "-framework QuartzCore" "-framework Metal" "-framework MetalKit")
