+++ src/emu/host/win/window.c
@@ -1,0 +1,66 @@
+/*
+ * Copyright (c) 2026 Rumbledethumps
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ *
+ * Windows window host: the Win32 WM resize seam and the sokol entry (window_run
+ * -> sapp_run with high_dpi for a native-resolution D3D11 backbuffer). The
+ * render/frame/present pipeline is in host/window_core.c.
+ */
+
+#ifndef WIN32_LEAN_AND_MEAN
+#define WIN32_LEAN_AND_MEAN
+#endif
+#ifndef NOMINMAX
+#define NOMINMAX
+#endif
+#include <windows.h>
+
+#include "emu/host/window.h"
+#include "emu/host/window_core.h"
+#include "sokol_app.h"
+#include "sokol_log.h"
+#include <stdint.h>
+
+void host_window_resize(int w, int h)
+{
+    HWND hwnd = (HWND)sapp_win32_get_hwnd();
+    if (!hwnd)
+        return;
+    /* w,h are client (== framebuffer/physical) px; grow by this window's DPI
+     * frame and keep the top-left corner. */
+    RECT r = {0, 0, w, h};
+    AdjustWindowRectExForDpi(&r,
+                             (DWORD)GetWindowLongPtrW(hwnd, GWL_STYLE), FALSE,
+                             (DWORD)GetWindowLongPtrW(hwnd, GWL_EXSTYLE),
+                             GetDpiForWindow(hwnd));
+    SetWindowPos(hwnd, NULL, 0, 0, r.right - r.left, r.bottom - r.top,
+                 SWP_NOMOVE | SWP_NOZORDER | SWP_NOACTIVATE);
+}
+
+void host_window_set_aspect_hint(int cw, int ch) { (void)cw, (void)ch; }
+void host_window_init(void) {}
+bool host_window_menu_active(void) { return false; }
+void host_window_menu_draw(void) {}
+
+int window_run(uint32_t *fb, double scale, bool have_scale, bool vsync, bool exit_on_halt)
+{
+    int win_w, win_h;
+    window_core_prepare(fb, scale, have_scale, vsync, exit_on_halt, &win_w, &win_h);
+    /* D3D11 leaves the backbuffer at LOGICAL size unless high_dpi is requested,
+     * so a DPI-scaled display DWM-stretches (smears) the menu/canvas; ask for a
+     * native-resolution backbuffer. */
+    sapp_run(&(sapp_desc){
+        .init_cb = window_core_init,
+        .frame_cb = window_core_frame,
+        .event_cb = window_core_event,
+        .cleanup_cb = window_core_cleanup,
+        .width = win_w,
+        .height = win_h,
+        .high_dpi = true,
+        .swap_interval = vsync ? 1 : 0, /* off: present uncapped (driver may ignore) */
+        .window_title = "Picocomputer 6502",
+        .logger.func = slog_func,
+    });
+    return 0;
+}
